refactor: separation between test file and remove duplicate test
